Extract JsonRpcServerBuilder
Macro doesn't play nice when chaining macro methods such as bindApi. This change is so that we can prevent chaining bindApi.
